ocr_service/requirements.txt
# OCR Service Dependencies
# Python 3.12+

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Image processing
Pillow==11.0.0
opencv-python-headless==4.10.0.84

# OCR
pytesseract==0.3.13

# Request handling
python-multipart==0.0.18

# Development & Testing
pytest==8.3.4
httpx==0.28.1

# Logging
python-json-logger==3.2.1
